+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 
 All notable changes to this project will be documented in this file.
+## Version [1.2.17] - Profile UI fixes.
+**P.R Raised by** : shubh404-SE
+**Date** : 2026-05-05
+
+### Changed
+  - ui them fixed for ui components and settings profile fields.
+  - bug fixed while entering or removing text from input fields of profile settings.
+---
 ## Version [1.2.16] - Announcement Feature
 **P.R Raised by** : shubh404-SE on **Date** : 2026-04-10